Added a sfd file for a dopushook.library.
This should match the hook interface. The contents will be mainly used to generate 68k->ppc cross calls. They are needed as the hook interface is 68k only. It probably would make sense to move these function into the dopus.library or into a dedicated library for modules, so this is likely a temporary solution for now.