Removed this file again.
The traps for the 68k emu are already present in callback_main.h, so there is no need to generate them from this file. If somebody want's to add the function into the dopus.library (or a new one), the file may be still useful. But it has to be recovered from the attic then.