Thanks to Roman for adding the OpenWorkbenchObject(A) patching code.
Following on from this, the patches now check to see if Workbench is available (by searching for its public port), and passing the call onto the original vectors. 

Currently, there is only debug to prove it works, but we will launch the requested objects ourselves if Workbench is not available (running in replacement mode).