Fixed some compiler warnings.
Added a TBGADGETOFFSET define, which, under AmigaOS4, spaces the iconify and padlock gadgets out so that the zoom gadget is not obstructed. While this appears to work with the standard GUI theme, it needs testing with other themes to verify the spacing. All other platforms specify the offset as ZERO, as before.