Revert "enable -mregparm=4"
This reverts commit 9dfab7d1c647b3594ace2db4cf81ebe4342a4735.

Some DOpus library functions (in modules) take parameters in a4 and a5.
This works against the usage of base-relative addressing.
In order to work around issues when calling such functions, some
of them (for instance (L_Config_Environment) Have been implemented
in assembly with two stub functions

One example is Config_Environment();

1) at the call site:

    The prototype is defined as:
    unsigned long Config_Environment(Cfg_Environment *,struct Screen *,UWORD *,IPCData *,IPCData *,UWORD,ULONG *,char *,Att_List *);
    There is no usual library inlining macro for Config_Environment, so al
    calls go through this prototype.
    Implemented as a stub inside Program/68k_asm_a4_a5.c in assembly.
    That assembly function expects the function parameters on the stack.
    It'll take the parameters from the stack, put them into registers and then
    do the actual library call.

2) at the callee site:
     (Modules/configopus/config_environment.c)
    _L_Config_Environment takes the two parameters in a4 and a5 and puts them
    into global variables, then forwards the call into 'ConfigEnv()' which in
    turn 'knows' that two of its parameters have been funneled through
    'global_main_ipc' and 'global_change_flags'

Now when -mregparms=4 was specified, the call site assembly stub did not account
for that and grabbed the parameters from the stack even though they were not there.
*boom*

In the long run, functions with an extremely long parameter list should be
revisited and use either structs or tag lists.
It also shows that DOpus might be able to compile with -fbaserel