Revert "Fixed standard drag routine so it no longer crashes when Dopus5"
This reverts commit 6c107500424364a8546d4cdf3678f583b4dea4da.

This fixes an issue with drag & drop where the dragged icon will just
be left wherever the drag ended (even if the drag was canceled), leaving
corrupted backdrops and windows behind.

The original commit moved the call to DrawDragList (and ultimately DrawGList)
_before_ the call to RemBob. But DrawGList needs to be called _after) RemBob
to fully remove the Bob from the screen.

The original commit was a fix for OS4, but here we only care about OS3,
so just revert it and be happy.