Fix custom dragging memory leak
When doing custom dragging, DOpus is abusing the GELs list for just keeping track of the
dragged images; not for actually drawing them.
It used to add the icons to the list using AddBob() (which seems to cause the additional
allocations that later leaked) in backdrop_show_drag(), but no symmetric RemBob would be
called in backdrop_stop_drag(). RemBob() would not work anyway, since it merely flags
the bob for removal, without performing the removal.
The actual removal is done via DrawGList(), which is never called for custom dragging.

Since for custom dragging all that matters is adding the object to the GELs list,
we can use AddVSprite() and RemVSprite() instead.
We make sure, RemVSprite is called after the call to DrawDragList(), so the dragged
image is properly removed from the screen before.